@fernandoportodev Task: Validar alterações no projeto

+++ src/main/java/org/generation/blog_pessoal/controller/PostagemController.java
@@ -1,6 +1,9 @@
 package org.generation.blog_pessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.blog_pessoal.model.Postagem;
 import org.generation.blog_pessoal.repository.PostagemRepository;
@@ -15,15 +18,17 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/postagens")
-@CrossOrigin("*")
+@CrossOrigin(origins = "*", allowedHeaders = "*")
 public class PostagemController {
 
-    @Autowired
-    private PostagemRepository repository;
+    
+    private @Autowired PostagemRepository repository;
 
     @GetMapping
     public ResponseEntity<List<Postagem>> getAll(){
@@ -32,7 +37,7 @@ public class PostagemController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Postagem> GetById(@PathVariable long id){
-        return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
+        return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
     }
 
     @GetMapping("/titulo/{titulo}")
@@ -41,17 +46,22 @@ public class PostagemController {
     }
     
     @PostMapping
-    public ResponseEntity<Postagem> post (@RequestBody Postagem postagem){
+    public ResponseEntity<Postagem> post (@Valid @RequestBody Postagem postagem){
         return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
     }
 
     @PutMapping
-    public ResponseEntity<Postagem> put (@RequestBody Postagem postagem){
-        return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
+    public ResponseEntity<Postagem> put (@Valid @RequestBody Postagem postagem){
+        return repository.findById(postagem.getId()).map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem))).orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
     }
 
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     public void delete(@PathVariable long id){
-        repository.deleteById(id);
-    }
+        Optional<Postagem> post = repository.findById(id);
+            if(post.isEmpty())
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
+            repository.deleteById(id);
+    }   
 }
