+++ src/main/java/org/generation/blog_pessoal/controller/TemaController.java
@@ -1,7 +1,9 @@
 package org.generation.blog_pessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
+import javax.validation.Valid;
 
 import org.generation.blog_pessoal.model.Tema;
 import org.generation.blog_pessoal.repository.TemaRepository;
@@ -16,15 +18,17 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 @RequestMapping("/tema")
 public class TemaController {
     
-    @Autowired
-    private TemaRepository repository;
+    
+    private @Autowired TemaRepository repository;
 
     @GetMapping
     public ResponseEntity<List<Tema>> getAll(){
@@ -36,23 +40,29 @@ public class TemaController {
         return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
     }
 
-    @GetMapping("/nome/{nome}")
-    public ResponseEntity<List<Tema>> getByName(String nome){
-        return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
+    @GetMapping("/descricao/{descricao}")
+    public ResponseEntity<List<Tema>> getByTitle(@PathVariable String descricao){
+        return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(descricao));
     }
 
     @PostMapping
-    public ResponseEntity<Tema> post (@RequestBody Tema tema){
+    public ResponseEntity<Tema> post (@Valid @RequestBody Tema tema){
         return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema));
     }
 
     @PutMapping
-    public ResponseEntity<Tema> put(@RequestBody Tema tema){
-        return ResponseEntity.ok(repository.save(tema));
+    public ResponseEntity<Tema> put(@Valid @RequestBody Tema tema){
+        return repository.findById(tema.getId()).map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema))).orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
     }
 
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     public void delete(@PathVariable Long id){
+        Optional<Tema> tema = repository.findById(id);
+
+        if(tema.isEmpty())
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            
         repository.deleteById(id);
     }
 }
